@@ -1,4 +1,4 @@
-class BlogPost < ActiveRecord::Base
+class BlogPost < ApplicationRecord
 
   has_and_belongs_to_many :topics, join_table:"blog_post_topics"
   has_many :footnotes
